updated per code review
