+++ cmd/notation/sign.go
@@ -9,6 +9,7 @@ import (
 	"github.com/notaryproject/notation-go"
 	"github.com/notaryproject/notation-go/crypto/timestamp"
 	"github.com/notaryproject/notation/internal/cmd"
+	"github.com/notaryproject/notation/internal/envelope"
 	"github.com/spf13/cobra"
 )
 
@@ -18,8 +19,6 @@ type signOpts struct {
 	timestamp       string
 	expiry          time.Duration
 	originReference string
-	push            bool
-	pushReference   string
 	pluginConfig    string
 	reference       string
 }
@@ -67,11 +66,8 @@ Example - Sign a container image using the image digest
 	cmd.SetPflagTimestamp(command.Flags(), &opts.timestamp)
 	cmd.SetPflagExpiry(command.Flags(), &opts.expiry)
 	cmd.SetPflagReference(command.Flags(), &opts.originReference)
-
-	command.Flags().BoolVar(&opts.push, "push", true, "push after successful signing")
-	command.Flags().StringVar(&opts.pushReference, "push-reference", "", "different remote to store signature")
-
 	cmd.SetPflagPluginConfig(command.Flags(), &opts.pluginConfig)
+
 	return command
 }
 
@@ -93,20 +89,16 @@ func runSign(command *cobra.Command, cmdOpts *signOpts) error {
 	}
 
 	// write out
-	if ref := cmdOpts.pushReference; cmdOpts.push && !(cmdOpts.Local && ref == "") {
-		if ref == "" {
-			ref = cmdOpts.reference
-		}
-		if _, err := pushSignature(command.Context(), &cmdOpts.SecureFlagOpts, ref, sig); err != nil {
-			return fmt.Errorf("fail to push signature to %q: %v: %v",
-				ref,
-				desc.Digest,
-				err,
-			)
-		}
+	ref := cmdOpts.reference
+	if _, err := pushSignature(command.Context(), &cmdOpts.SecureFlagOpts, ref, sig); err != nil {
+		return fmt.Errorf("fail to push signature to %q: %v: %v",
+			ref,
+			desc.Digest,
+			err,
+		)
 	}
 
-	fmt.Println("Sign succeeded for", desc.Digest)
+	fmt.Println(desc.Digest)
 	return nil
 }
 
@@ -135,4 +127,29 @@ func prepareSigningContent(ctx context.Context, opts *signOpts) (notation.Descri
 		TSA:          tsa,
 		PluginConfig: pluginConfig,
 	}, nil
+}
+
+func pushSignature(ctx context.Context, opts *SecureFlagOpts, ref string, sig []byte) (notation.Descriptor, error) {
+	// initialize
+	sigRepo, err := getSignatureRepository(opts, ref)
+	if err != nil {
+		return notation.Descriptor{}, err
+	}
+	manifestDesc, err := getManifestDescriptorFromReference(ctx, opts, ref)
+	if err != nil {
+		return notation.Descriptor{}, err
+	}
+
+	// core process
+	// pass in nonempty annotations if needed
+	sigMediaType, err := envelope.SpeculateSignatureEnvelopeFormat(sig)
+	if err != nil {
+		return notation.Descriptor{}, err
+	}
+	sigDesc, _, err := sigRepo.PutSignatureManifest(ctx, sig, sigMediaType, manifestDesc, make(map[string]string))
+	if err != nil {
+		return notation.Descriptor{}, fmt.Errorf("put signature manifest failure: %v", err)
+	}
+
+	return sigDesc, nil
 }
